the SHA1 hash needs to be 5 chunks 32 bits each

+++ rutil/Sha1.hxx
@@ -28,6 +28,7 @@
 #include <iostream>
 #include <string>
 #include "Data.hxx"
+#include <stdint.h>
 
 namespace resip
 {
@@ -43,13 +44,14 @@ public:
     static std::string from_file(const std::string &filename);
 
 private:
-    typedef unsigned long int uint32;   /* just needs to be at least 32bit */
-    typedef unsigned long long uint64;  /* just needs to be at least 64bit */
+    typedef uint32_t uint32;
+    typedef uint64_t uint64;
 
     static const unsigned int DIGEST_INTS = 5;  /* number of 32bit integers per SHA1 digest */
     static const unsigned int BLOCK_INTS = 16;  /* number of 32bit integers per SHA1 block */
     static const unsigned int BLOCK_BYTES = BLOCK_INTS * 4;
 
+    // the SHA1 digest is 160 bits split into 5 chunks, each chunk needs to be exactly 32 bits
     uint32 digest[DIGEST_INTS];
     std::string buffer;
     uint64 transforms;
